BackEnd/requirements.txt
cryptography==3.1.1
Django==3.1.4
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-social-oauth2==1.0.8
Pillow==8.0.1
uritemplate==3.0.1
social-auth-app-django==4.0.0
social-auth-core==3.3.3
coreapi==2.3.3